fix: Migrate logout logic to plugins
related #507
